Refactor setup controller and registry manipulator
